completed the page 0.0.0.1

+++ src/App.jsx
@@ -7,10 +7,12 @@ import FormComponent from './components/FormComponent';
 import CardComponent from './components/CardComponent';
 import BlogDetail from './components/BlogDetail';
 import Mainpage from './components/Mainpage';
+import Swal from 'sweetalert2';
 
 function App() {
   const [blogs, setBlogs] = useState([]);
   const [isFormOpen, setIsFormOpen] = useState(false);
+  const [blogToEdit, setBlogToEdit] = useState(null);
 
   useEffect(() => {
     const existingBlogs = JSON.parse(localStorage.getItem('blogs')) || [];
@@ -18,32 +20,62 @@ function App() {
   }, []);
 
   const handleDelete = (index) => {
-    const updatedBlogs = blogs.filter((_, i) => i !== index);
-    setBlogs(updatedBlogs);
-    localStorage.setItem('blogs', JSON.stringify(updatedBlogs));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const updatedBlogs = blogs.filter((_, i) => i !== index);
+        setBlogs(updatedBlogs);
+        localStorage.setItem('blogs', JSON.stringify(updatedBlogs));
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
+        });
+      }
+    });
   };
 
   const handleCreate = () => {
-    setIsFormOpen(true); 
+    setBlogToEdit(null);
+    setIsFormOpen(true);
+  };
+
+  const handleEdit = (index) => {
+    setBlogToEdit({ ...blogs[index], index });
+    setIsFormOpen(true);
   };
 
   const handleCloseForm = () => {
-    setIsFormOpen(false); 
+    setIsFormOpen(false);
   };
 
   const handleSubmit = (newBlog) => {
-    setBlogs((prevBlogs) => [...prevBlogs, newBlog]);
-    localStorage.setItem('blogs', JSON.stringify([...blogs, newBlog]));
+    if (blogToEdit) {
+      const updatedBlogs = blogs.map((blog, i) =>
+        i === blogToEdit.index ? newBlog : blog
+      );
+      setBlogs(updatedBlogs);
+      localStorage.setItem('blogs', JSON.stringify(updatedBlogs));
+    } else {
+      setBlogs((prevBlogs) => [...prevBlogs, newBlog]);
+      localStorage.setItem('blogs', JSON.stringify([...blogs, newBlog]));
+    }
   };
 
   return (
     <BrowserRouter>
-      <Navbar onCreate={handleCreate} /> 
+      <Navbar onCreate={handleCreate} />
       <div className="flex gap-0 p-0">
         <Menu />
         <div>
           <Routes>
-           
             <Route
               path="/"
               element={
@@ -53,25 +85,22 @@ function App() {
                 </>
               }
             />
-
-            
             <Route
               path="/blogs"
               element={
-                <div className="flex flex-wrap gap-4 mx-8 justify-center  md:grid-cols-3 py-6 ">
+                <div className="flex flex-wrap gap-6 mx-8 justify-center md:grid-cols-3 py-6">
                   {blogs.map((blog, index) => (
                     <CardComponent
                       key={index}
                       index={index}
                       blog={blog}
                       onDelete={() => handleDelete(index)}
+                      onEdit={() => handleEdit(index)}
                     />
                   ))}
                 </div>
               }
             />
-
-           
             <Route
               path="/blogs/:index"
               element={<BlogDetail blogs={blogs} onDelete={handleDelete} />}
@@ -79,13 +108,12 @@ function App() {
           </Routes>
         </div>
       </div>
-
-      
       {isFormOpen && (
         <FormComponent
           isOpen={isFormOpen}
           onClose={handleCloseForm}
           onSubmit={handleSubmit}
+          blogToEdit={blogToEdit}
         />
       )}
     </BrowserRouter>
